// src/menu-drawer/types.ts
export type MenuItemProps = {
    /**
     * Иконка элемента меню.
     *
     * Ожидается, что это будет React-элемент, отображаемый рядом с текстом.
     */
    icon: React.ReactNode
    /**
     * Текст элемента меню.
     *
     * Отображаемый заголовок или название пункта меню.
     */
    text: string
    /**
     * Обработчик клика по элементу.
     *
     * Функция, вызываемая при нажатии на элемент меню.
     */
    onClick?: () => void
    /**
     * Обработчик события наведения.
     *
     * Функция, вызываемая при наведении на элемент меню.
     */
    onHover?: () => void
    /**
     * Отключение элемента меню.
     *
     * Делает элемент неактивным и недоступным для взаимодействия.
     */
    disabled?: boolean
    /**
     * Скрытие элемента меню.
     *
     * Определяет, будет ли элемент отображаться в меню или нет.
     */
    hidden?: boolean
}

export type MenuDrawerProps = {
    /**
     * Список элементов меню.
     *
     * Массив объектов с параметрами для каждого элемента меню.
     */
    items: MenuItemProps[]
    /**
     * Индекс активного элемента меню.
     *
     * Указывает на текущий выбранный или активный пункт меню.
     */
    activeItem?: number
    /**
     * Обработчик событий клавиатуры.
     *
     * Функция, вызываемая при нажатии клавиш, когда меню в фокусе.
     */
    onKeyDown?: (e: React.KeyboardEvent) => void
}